# docs/contracts/oneshim-web.v1.openapi.yaml
openapi: 3.0.3
info:
  title: ONESHIM Local Web API
  version: "v1"
  description: |
    Auto-generated from docs/contracts/http-interface-manifest.v1.json.
    This snapshot is transport-level and contract-focused.
  x-oneshim-document-version: 3
  x-oneshim-updated-at: "2026-03-20"
  x-oneshim-routes-file: "crates/oneshim-web/src/routes.rs"
  x-oneshim-contracts-crate: "crates/oneshim-api-contracts"
servers:
  - url: /
paths:
...
  "/api/pomodoro/cancel":
    post:
      tags:
        - pomodoro
      operationId: pomodoro_post_api_pomodoro_cancel
      summary: "POST /api/pomodoro/cancel"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/pomodoro/complete":
    post:
      tags:
        - pomodoro
      operationId: pomodoro_post_api_pomodoro_complete
      summary: "POST /api/pomodoro/complete"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/pomodoro/current":
    get:
      tags:
        - pomodoro
      operationId: pomodoro_get_api_pomodoro_current
      summary: "GET /api/pomodoro/current"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/pomodoro/start":
    post:
      tags:
        - pomodoro
      operationId: pomodoro_post_api_pomodoro_start
      summary: "POST /api/pomodoro/start"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
        "default":
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: true